src/Services/transcript/requirements.txt
Flask
dapr
azure-cognitiveservices-speech
